Search functionality
Added:
search functionality to filter through on a array of dummy data
populated the dashboard.tsx with the dummy data.
made a simple dashboard (hovedlisten)
Made detailview which presents only the information conserning the searched word. NOT FINISHED.

Next step:
The search results rendes a list of shipments, make a new component or a page that displays the detailed view of the wanted shipement. Might have to redo some things..e